ARROW-8658: [C++][Dataset] Move Forest to arrow/dataset
