+++ cpp/src/arrow/CMakeLists.txt
@@ -418,7 +418,6 @@ if(ARROW_FILESYSTEM)
               filesystem/filesystem.cc
               filesystem/localfs.cc
               filesystem/mockfs.cc
-              filesystem/path_forest.cc
               filesystem/path_util.cc
               filesystem/util_internal.cc)
